refactoring some part of code

+++ core/src/com/mygdx/game/Cha.java
@@ -3,7 +3,7 @@ package com.mygdx.game;
 import com.badlogic.gdx.math.Vector2;
 
 public class Cha extends GameObject{
-	public Cha(int x, int y){
+	public Cha(float x, float y){
 		super(x, y);
 	}
